Rename UEntityBuilder to UEntityFactory
