+++ src/test/java/org/eclipse/uprotocol/uri/serializer/MicroUriSerializerTest.java
@@ -24,7 +24,8 @@
 package org.eclipse.uprotocol.uri.serializer;
 
 import com.google.protobuf.ByteString;
-import org.eclipse.uprotocol.uri.builder.UResourceBuilder;
+
+import org.eclipse.uprotocol.uri.factory.UResourceBuilder;
 import org.eclipse.uprotocol.uri.validator.UriValidator;
 import org.eclipse.uprotocol.v1.UAuthority;
 import org.eclipse.uprotocol.v1.UEntity;
