+++ src/test/java/org/eclipse/uprotocol/uri/validator/UriValidatorTest.java
@@ -25,7 +25,7 @@
 package org.eclipse.uprotocol.uri.validator;
 
 
-import org.eclipse.uprotocol.uri.builder.UResourceBuilder;
+import org.eclipse.uprotocol.uri.factory.UResourceBuilder;
 import org.eclipse.uprotocol.uri.serializer.LongUriSerializer;
 import org.eclipse.uprotocol.v1.UAuthority;
 import org.eclipse.uprotocol.v1.UEntity;
